+++ examples/data-sources/nxos_port_channel_interface/data-source.tf
@@ -1,0 +1,3 @@
+data "nxos_port_channel_interface" "example" {
+  interface_id = "po1"
+}