+++ examples/data-sources/nxos_vpc_interface/data-source.tf
@@ -1,0 +1,3 @@
+data "nxos_vpc_interface" "example" {
+  vpc_interface_id = 1
+}