+++ examples/resources/nxos_port_channel_interface/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_port_channel_interface.example "sys/intf/aggr-[po1]"