+++ examples/resources/nxos_vpc_interface/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_vpc_interface.example "sys/vpc/inst/dom/if-[1]"