+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - BREAKING CHANGE: Remove  `nxos_ipv4_access_list_policy_egress_interface_instance` resource and data source
 - Add `nxos_vpc_instance` resource and data source
 - Add `nxos_vpc_domain` resource and data source
+- Add `nxos_vpc_interface` resource and data source
+- Add `nxos_port_channel_interface` resource and data source
 
 ## 0.3.23