+++ docs/guides/changelog.md
@@ -16,6 +16,8 @@ description: |-
 - BREAKING CHANGE: Remove  `nxos_ipv4_access_list_policy_egress_interface_instance` resource and data source
 - Add `nxos_vpc_instance` resource and data source
 - Add `nxos_vpc_domain` resource and data source
+- Add `nxos_vpc_interface` resource and data source
+- Add `nxos_port_channel_interface` resource and data source
 
 ## 0.3.23